// .devops/aws/src/imports/k8s.d.ts
export declare class KubeSecret extends ApiObject {
  /**
   * Returns the apiVersion and kind for "io.k8s.api.core.v1.Secret"
   */
  static readonly GVK: GroupVersionKind;
  /**
   * Renders a Kubernetes manifest for "io.k8s.api.core.v1.Secret".
   *
   * This can be used to inline resource manifests inside other objects (e.g. as templates).
   *
   * @param props initialization props
   */
  static manifest(props?: KubeSecretProps): any;
  /**
   * Defines a "io.k8s.api.core.v1.Secret" API object
   * @param scope the scope in which to define this object
   * @param id a scope-local name for the object
   * @param props initialization props
   */
  constructor(scope: Construct, id: string, props?: KubeSecretProps);
  /**
   * Renders the object to Kubernetes JSON.
   */
  toJson(): any;
}
// ...
export declare class KubeRole extends ApiObject {
  /**
   * Returns the apiVersion and kind for "io.k8s.api.rbac.v1.Role"
   */
  static readonly GVK: GroupVersionKind;
  /**
   * Renders a Kubernetes manifest for "io.k8s.api.rbac.v1.Role".
   *
   * This can be used to inline resource manifests inside other objects (e.g. as templates).
   *
   * @param props initialization props
   */
  static manifest(props?: KubeRoleProps): any;
  /**
   * Defines a "io.k8s.api.rbac.v1.Role" API object
   * @param scope the scope in which to define this object
   * @param id a scope-local name for the object
   * @param props initialization props
   */
  constructor(scope: Construct, id: string, props?: KubeRoleProps);
  /**
   * Renders the object to Kubernetes JSON.
   */
  toJson(): any;
}
/**
 * RoleBinding references a role, but does not contain it.  It can reference a Role in the same namespace or a ClusterRole in the global namespace. It adds who information via Subjects and namespace information by which namespace it exists in.  RoleBindings in a given namespace only have effect in that namespace.
 *
 * @schema io.k8s.api.rbac.v1.RoleBinding
 */
export declare class KubeRoleBinding extends ApiObject {
  /**
   * Returns the apiVersion and kind for "io.k8s.api.rbac.v1.RoleBinding"
   */
  static readonly GVK: GroupVersionKind;
  /**
   * Renders a Kubernetes manifest for "io.k8s.api.rbac.v1.RoleBinding".
   *
   * This can be used to inline resource manifests inside other objects (e.g. as templates).
   *
   * @param props initialization props
   */
  static manifest(props: KubeRoleBindingProps): any;
  /**
   * Defines a "io.k8s.api.rbac.v1.RoleBinding" API object
   * @param scope the scope in which to define this object
   * @param id a scope-local name for the object
   * @param props initialization props
   */
  constructor(scope: Construct, id: string, props: KubeRoleBindingProps);
  /**
   * Renders the object to Kubernetes JSON.
   */
  toJson(): any;
}
// ...
export interface KubeSecretProps {
  /**
   * Data contains the secret data. Each key must consist of alphanumeric characters, '-', '_' or '.'. The serialized form of the secret data is a base64 encoded string, representing the arbitrary (possibly non-string) data value here. Described in https://tools.ietf.org/html/rfc4648#section-4
   *
   * @schema io.k8s.api.core.v1.Secret#data
   */
  readonly data?: {
    [key: string]: string;
  };
  /**
   * Immutable, if set to true, ensures that data stored in the Secret cannot be updated (only object metadata can be modified). If not set to true, the field can be modified at any time. Defaulted to nil.
   *
   * @schema io.k8s.api.core.v1.Secret#immutable
   */
  readonly immutable?: boolean;
  /**
   * Standard object's metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
   *
   * @schema io.k8s.api.core.v1.Secret#metadata
   */
  readonly metadata?: ObjectMeta;
  /**
   * stringData allows specifying non-binary secret data in string form. It is provided as a write-only input field for convenience. All keys and values are merged into the data field on write, overwriting any existing values. The stringData field is never output when reading from the API.
   *
   * @schema io.k8s.api.core.v1.Secret#stringData
   */
  readonly stringData?: {
    [key: string]: string;
  };
  /**
   * Used to facilitate programmatic handling of secret data.
   *
   * @schema io.k8s.api.core.v1.Secret#type
   */
  readonly type?: string;
}
// ...
export interface KubeRoleProps {
  /**
   * Standard object's metadata.
   *
   * @schema io.k8s.api.rbac.v1.Role#metadata
   */
  readonly metadata?: ObjectMeta;
  /**
   * Rules holds all the PolicyRules for this Role
   *
   * @schema io.k8s.api.rbac.v1.Role#rules
   */
  readonly rules?: PolicyRule[];
}
// ...
export interface KubeRoleBindingProps {
  /**
   * Standard object's metadata.
   *
   * @schema io.k8s.api.rbac.v1.RoleBinding#metadata
   */
  readonly metadata?: ObjectMeta;
  /**
   * RoleRef can reference a Role in the current namespace or a ClusterRole in the global namespace. If the RoleRef cannot be resolved, the Authorizer must return an error.
   *
   * @schema io.k8s.api.rbac.v1.RoleBinding#roleRef
   */
  readonly roleRef: RoleRef;
  /**
   * Subjects holds references to the objects the role applies to.
   *
   * @schema io.k8s.api.rbac.v1.RoleBinding#subjects
   */
  readonly subjects?: Subject[];
}
// ...
export interface PolicyRule {
  /**
   * APIGroups is the name of the APIGroup that contains the resources.  If multiple API groups are specified, any action requested against one of the enumerated resources in any API group will be allowed.
   *
   * @schema io.k8s.api.rbac.v1.PolicyRule#apiGroups
   */
  readonly apiGroups?: string[];
  /**
   * NonResourceURLs is a set of partial urls that a user should have access to.  *s are allowed, but only as the full, final step in the path Since non-resource URLs are not namespaced, this field is only applicable for ClusterRoles referenced from a ClusterRoleBinding. Rules can either apply to API resources (such as "pods" or "secrets") or non-resource URL paths (such as "/api"),  but not both.
   *
   * @schema io.k8s.api.rbac.v1.PolicyRule#nonResourceURLs
   */
  readonly nonResourceUrLs?: string[];
  /**
   * ResourceNames is an optional white list of names that the rule applies to.  An empty set means that everything is allowed.
   *
   * @schema io.k8s.api.rbac.v1.PolicyRule#resourceNames
   */
  readonly resourceNames?: string[];
  /**
   * Resources is a list of resources this rule applies to. '*' represents all resources.
   *
   * @schema io.k8s.api.rbac.v1.PolicyRule#resources
   */
  readonly resources?: string[];
  /**
   * Verbs is a list of Verbs that apply to ALL the ResourceKinds and AttributeRestrictions contained in this rule. '*' represents all verbs.
   *
   * @schema io.k8s.api.rbac.v1.PolicyRule#verbs
   */
  readonly verbs: string[];
}
/**
 * Converts an object of type 'PolicyRule' to JSON representation.
 */
export declare function toJson_PolicyRule(
  obj: PolicyRule | undefined,
): Record<string, any> | undefined;
/**
 * RoleRef contains information that points to the role being used
 *
 * @schema io.k8s.api.rbac.v1.RoleRef
 */
export interface RoleRef {
  /**
   * APIGroup is the group for the resource being referenced
   *
   * @schema io.k8s.api.rbac.v1.RoleRef#apiGroup
   */
  readonly apiGroup: string;
  /**
   * Kind is the type of resource being referenced
   *
   * @schema io.k8s.api.rbac.v1.RoleRef#kind
   */
  readonly kind: string;
  /**
   * Name is the name of resource being referenced
   *
   * @schema io.k8s.api.rbac.v1.RoleRef#name
   */
  readonly name: string;
}
// ...
export interface Subject {
  /**
   * APIGroup holds the API group of the referenced subject. Defaults to "" for ServiceAccount subjects. Defaults to "rbac.authorization.k8s.io" for User and Group subjects.
   *
   * @default for ServiceAccount subjects. Defaults to "rbac.authorization.k8s.io" for User and Group subjects.
   * @schema io.k8s.api.rbac.v1.Subject#apiGroup
   */
  readonly apiGroup?: string;
  /**
   * Kind of object being referenced. Values defined by this API group are "User", "Group", and "ServiceAccount". If the Authorizer does not recognized the kind value, the Authorizer should report an error.
   *
   * @schema io.k8s.api.rbac.v1.Subject#kind
   */
  readonly kind: string;
  /**
   * Name of the object being referenced.
   *
   * @schema io.k8s.api.rbac.v1.Subject#name
   */
  readonly name: string;
  /**
   * Namespace of the referenced object.  If the object kind is non-namespace, such as "User" or "Group", and this value is not empty the Authorizer should report an error.
   *
   * @schema io.k8s.api.rbac.v1.Subject#namespace
   */
  readonly namespace?: string;
}